use int32 playerid (#613)
* use int32 if possible

* remove include in effectset.h

* use int32 playerid

* update playerid in class effect

* update playerid in class field

+++ effect.cpp
@@ -657,7 +657,7 @@ int32 effect::reset(uint32 reset_level, uint32 reset_type) {
 	}
 	return FALSE;
 }
-void effect::dec_count(uint32 playerid) {
+void effect::dec_count(uint8 playerid) {
 	if(!is_flag(EFFECT_FLAG_COUNT_LIMIT))
 		return;
 	if(count_limit == 0)
